split out server and devserver
