+++ package.json
@@ -78,7 +78,7 @@
     "start": "npm-run-all --parallel generate:first-styles generate:variables --serial start:react",
     "build": "npm-run-all --parallel build:server build:frontend",
     "build:frontend": "npm-run-all --parallel generate:first-styles generate:variables --serial build:react",
-    "start:dev-server": "nodemon --exec ts-node --project server/tsconfig.json  server/server.ts",
+    "start:dev-server": "nodemon --exec ts-node --project server/tsconfig.json  server/dev-server.ts",
     "start:server": "pm2 start ecosystem.config.js --env production",
     "start:react": "react-scripts start",
     "build:react": "react-app-rewired build",
